Refactor MainActivity runtime tuning into separate file
